consolidate Mathquill Field types
