+++ packages/math-input/src/components/key-handlers/handle-arrow.ts
@@ -1,9 +1,9 @@
+import {MathFieldActionType} from "../../types";
 import {
     maybeFindCommand,
     maybeFindCommandBeforeParens,
 } from "../input/mathquill-helpers";
 import {mathQuillInstance} from "../input/mathquill-instance";
-import {MathFieldActionType} from "../input/mathquill-types";
 
 import type Key from "../../data/keys";
 import type {MathFieldInterface} from "../input/mathquill-types";
